requirements.txt
pandas>=2.0
openpyxl>=3.1        # Excel read/write support for pandas
requests>=2.31
PyYAML>=6.0

